fs: Workaround the compiler's bad optimization
When compiling the kernel with  the compiler
"aarch64-linux-android-gcc (GCC) 4.9.x-google 20140827 (prerelease)"
The compiler seems to be optimizing a value out in the do_sync_write
function and is incorrectly passing the pointer instead of the
dereferenced value.
Force the compiler to use the right value by passing the dereferenced
pointer instead of another variable that is assigned with the dereferenced
value.

Change-Id: I9d54dd7c9165a495073dc38227bd4e2144463b70
Signed-off-by: Nikhilesh Reddy <reddyn@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>